refactor(frontend): nest graduation reports, create components dir
