+++ frontend/src/app/(scoreboard)/graduations/reports/[cohort]/page.tsx
@@ -1,5 +1,5 @@
 import { listGraduationsByCohort } from '@/api/graduationApi';
-import GraduationReportsGrid from '@/graduation/RecentGraduationReports';
+import GraduationReportsGrid from '@/components/graduations/GraduationReportsGrid';
 
 interface PageProps {
     params: {
