+++ frontend/src/app/(scoreboard)/graduations/reports/page.tsx
@@ -1,5 +1,5 @@
 import { listGraduationsByDate } from '@/api/graduationApi';
-import GraduationReportsGrid from '@/graduation/RecentGraduationReports';
+import GraduationReportsGrid from '@/components/graduations/GraduationReportsGrid';
 
 export default async function Page() {
     const graduations = await listGraduationsByDate();
