+++ frontend/src/components/graduations/GraduationLinkCard.tsx
