+++ frontend/src/components/graduations/GraduationReport.tsx
